reorganize kes override to make it more reusable for integration testing
