+++ packages/deployment/.gitignore
@@ -1,0 +1,1 @@
+app/*.js
